requirements.txt
pytest==7.0.1
pytest-bdd==5.0.0
selenium==4.1.1
webdriver-manager==3.5.3
browserstack-local==1.2.3